Synced bullet count with weapon
