+++ Assets/Materials/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7cf784623e21d144f9f4ea675c50e83e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
