+++ Assets/Prefabs/Weapon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1654fc42656b14b4b88e8ca4ff0b2202
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
