+++ Assets/Prefabs/WeaponSupplier.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1bfa3628f34d84148bef0086d44bcdb1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
