+++ Assets/Scripts/WeaponSupplier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74114460f99189147823646587e7579d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
